+++ app/grails-app/controllers/com/k_int/kbplus/MyInstitutionController.groovy
@@ -1,8 +1,11 @@
 package com.k_int.kbplus
 
 import com.k_int.kbplus.auth.User
+import com.k_int.kbplus.auth.UserOrg
+import de.laser.ContextService
 import de.laser.helper.DebugAnnotation
 import grails.converters.JSON
+import grails.plugin.springsecurity.SpringSecurityUtils
 import grails.plugin.springsecurity.annotation.Secured
 import org.apache.poi.hslf.model.*
 import org.apache.poi.hssf.usermodel.*
@@ -30,11 +33,10 @@ class MyInstitutionController {
     def institutionsService
     def docstoreService
     def tsvSuperlifterService
-    def permissionHelperService
+    def accessService
     def contextService
     def taskService
     def filterService
-    def accessService
 
     // copied from
     static String INSTITUTIONAL_LICENSES_QUERY      = " from License as l where exists ( select ol from OrgRole as ol where ol.lic = l AND ol.org = :lic_org and ol.roleType = :org_role ) AND (l.status!=:lic_status or l.status=null ) "
@@ -54,24 +56,24 @@ class MyInstitutionController {
             new SimpleDateFormat('yyyy')
     ];
 
-    @DebugAnnotation(test='hasAffiliation("INST_USER")')
-    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_USER") })
+    @DebugAnnotation(test='hasAffiliation("INST_ADM")')
+    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_ADM") })
     def index() {
         // Work out what orgs this user has admin level access to
         def result = [:]
+        result.institution  = contextService.getOrg()
         result.user = User.get(springSecurityService.principal.id)
+        def currentOrg = contextService.getOrg()
         log.debug("index for user with id ${springSecurityService.principal.id} :: ${result.user}");
 
         if ( result.user ) {
           result.userAlerts = alertsService.getAllVisibleAlerts(result.user);
-          result.staticAlerts = alertsService.getStaticAlerts(request);
-
+          //result.staticAlerts = alertsService.getStaticAlerts(request);
 
           if ((result.user.affiliations == null) || (result.user.affiliations.size() == 0)) {
               redirect controller: 'profile', action: 'index'
           } else {
           }
-
         }
         else {
           log.error("Failed to find user in database");
@@ -95,7 +97,7 @@ class MyInstitutionController {
         def (tip_property, property_field) = (params.sort ?: 'title-title').split("-")
         def list_order = params.order ?: 'asc'
 
-        if (current_inst && ! permissionHelperService.checkUserIsMember(result.user, current_inst)) {
+        if (current_inst && ! accessService.checkUserIsMember(result.user, current_inst)) {
             flash.error = message(code:'myinst.error.noMember', args:[current_inst.name]);
             response.sendError(401)
             return;
@@ -136,7 +138,7 @@ class MyInstitutionController {
 
         result.tips = results
         result.institution = current_inst
-        result.editable = permissionHelperService.hasUserWithRole(result.user, current_inst, 'INST_ADM')
+        result.editable = accessService.checkUserOrgRole(result.user, current_inst, 'INST_ADM')
         result
     }
 
@@ -157,6 +159,18 @@ class MyInstitutionController {
         result
     }
 
+    @DebugAnnotation(test='hasAffiliation("INST_ADM")')
+    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_ADM") })
+    def manageAffiliationRequests() {
+        def result = [:]
+        result.institution        = contextService.getOrg()
+        result.user               = User.get(springSecurityService.principal.id)
+        result.editable           = true // inherit
+        result.pendingRequestsOrg = UserOrg.findAllByStatusAndOrg(UserOrg.STATUS_PENDING, contextService.getOrg(), [sort:'dateRequested'])
+
+        result
+    }
+
     @DebugAnnotation(test='hasAffiliation("INST_USER")')
     @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_USER") })
     def actionLicenses() {
@@ -174,13 +188,7 @@ class MyInstitutionController {
         def result = setResultGenerics()
         result.transforms = grailsApplication.config.licenseTransforms
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
-            flash.error = message(code:'myinst.error.noMember', args:[result.institution.name]);
-            response.sendError(401)
-            return;
-        }
-
-        result.is_inst_admin = permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM')
+        result.is_inst_admin = accessService.checkUserOrgRole(result.user, result.institution, 'INST_ADM')
 
         def date_restriction = null;
         def sdf = new java.text.SimpleDateFormat(message(code:'default.date.format.notime', default:'yyyy-MM-dd'))
@@ -341,13 +349,13 @@ class MyInstitutionController {
         result.max = params.max ? Integer.parseInt(params.max) : result.user.defaultPageSize;
         result.offset = params.offset ? Integer.parseInt(params.offset) : 0;
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (! accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = message(code:'myinst.error.noMember', args:[result.institution.name]);
             response.sendError(401)
             return;
         }
 
-        result.is_inst_admin = permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM')
+        result.is_inst_admin = accessService.checkUserOrgRole(result.user, result.institution, 'INST_ADM')
 
         def template_license_type = RefdataCategory.lookupOrCreate('License Type', 'Template');
         def qparams = [template_license_type]
@@ -441,13 +449,7 @@ class MyInstitutionController {
         }
         */
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
-            flash.error = message(code:'myinst.currentSubscriptions.no_permission', args:[result.institution.name]);
-            response.sendError(401)
-            return;
-        }
-
-        result.editable = permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM')
+        result.editable = accessService.checkUserOrgRole(result.user, result.institution, 'INST_ADM')
 
         def role_sub = RefdataCategory.lookupOrCreate('Organisational Role', 'Subscriber');
         def role_sub_consortia = RefdataCategory.lookupOrCreate('Organisational Role', 'Subscription Consortia');
@@ -566,7 +568,7 @@ from Subscription as s where (
         }
 
         // if ( ! permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM') ) {
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (! accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = message(code:'myinst.currentSubscriptions.no_permission', args:[result.institution.name]);
             response.sendError(401)
             result.is_inst_admin = false;
@@ -621,7 +623,7 @@ from Subscription as s where (
         def result = setResultGenerics()
         result.orgType = result.institution.orgType
         
-        result.editable = permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM')
+        result.editable = accessService.checkUserOrgRole(result.user, result.institution, 'INST_ADM')
 
         if (result.editable) {
             def cal = new java.util.GregorianCalendar()
@@ -666,7 +668,7 @@ from Subscription as s where (
           orgRole = role_sub
         }
 
-        if (permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM')) {
+        if (accessService.checkUserOrgRole(result.user, result.institution, 'INST_ADM')) {
 
             def sdf = new SimpleDateFormat(message(code:'default.date.format.notime', default:'yyyy-MM-dd'))
             def startDate = sdf.parse(params.valid_from)
@@ -807,7 +809,7 @@ from Subscription as s where (
         def user = User.get(springSecurityService.principal.id)
         def org = contextService.getOrg()
 
-        if (! permissionHelperService.hasUserWithRole(user, org, 'INST_ADM')) {
+        if (! accessService.checkUserOrgRole(user, org, 'INST_ADM')) {
             flash.error = message(code:'myinst.error.noAdmin', args:[org.name]);
             response.sendError(401)
             // render(status: '401', text:"You do not have permission to access ${org.name}. Please request access on the profile page");
@@ -838,7 +840,7 @@ from Subscription as s where (
         def user = User.get(springSecurityService.principal.id)
         def org = contextService.getOrg()
 
-        if (! permissionHelperService.hasUserWithRole(user, org, 'INST_ADM')) {
+        if (! accessService.checkUserOrgRole(user, org, 'INST_ADM')) {
             flash.error = message(code:'myinst.error.noAdmin', args:[org.name]);
             response.sendError(401)
             // render(status: '401', text:"You do not have permission to access ${org.name}. Please request access on the profile page");
@@ -870,7 +872,7 @@ from Subscription as s where (
         log.debug("deleteLicense ${params}");
         def result = setResultGenerics()
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (! accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = message(code:'myinst.error.noMember', args:[result.institution.name]);
             response.sendError(401)
             // render(status: '401', text:"You do not have permission to access ${result.institution.name}. Please request access on the profile page");
@@ -930,7 +932,7 @@ from Subscription as s where (
         def user = User.get(springSecurityService.principal.id)
         def institution = contextService.getOrg()
 
-        if (! permissionHelperService.checkUserIsMember(user, institution)) {
+        if (! accessService.checkUserIsMember(user, institution)) {
             flash.error = message(code:'myinst.error.noMember', args:[result.institution.name]);
             response.sendError(401)
             // render(status: '401', text:"You do not have permission to access ${institution.name}. Please request access on the profile page");
@@ -989,12 +991,6 @@ from Subscription as s where (
         
         log.debug("Viable Org-IDs are: ${viableOrgs}, active Inst is: ${result.institution.id}")
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
-            flash.error = message(code:'myinst.error.noMember', args:[result.institution.name]);
-            response.sendError(401)
-            return;
-        }
-
         // Set Date Restriction
         def date_restriction = null;
 
@@ -1012,7 +1008,7 @@ from Subscription as s where (
         }
         
         // Set is_inst_admin
-        result.is_inst_admin = permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM')
+        result.is_inst_admin = accessService.checkUserOrgRole(result.user, result.institution, 'INST_ADM')
 
         // Set offset and max
         result.max = params.max ? Integer.parseInt(params.max) : result.user.defaultPageSize;
@@ -1334,7 +1330,7 @@ AND EXISTS (
         def user = User.get(springSecurityService.principal.id)
         def institution = contextService.getOrg()
 
-        if (! permissionHelperService.checkUserIsMember(user, institution)) {
+        if (! accessService.checkUserIsMember(user, institution)) {
             flash.error = message(code:'myinst.error.noMember', args:[institution.name]);
             response.sendError(401)
             // render(status: '401', text:"You do not have permission to access ${institution.name}. Please request access on the profile page");
@@ -1411,7 +1407,7 @@ AND EXISTS (
         result.institution = contextService.getOrg()
         result.user = springSecurityService.getCurrentUser()
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (! accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = message(code:'myinst.error.noMember', args:[(result?.institution?.name?: message(code:'myinst.error.noMember.ph', default:'the selected institution'))]);
             response.sendError(401)
             // render(status: '401', text:"You do not have permission to access ${result.institution.name}. Please request access on the profile page");
@@ -2138,11 +2134,11 @@ AND EXISTS (
     def renewalsnoPackageChange() {
         def result = setResultGenerics()
 
-        if (!permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (!accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = message(code: 'myinst.error.noMember', args: [result.institution.name]);
             response.sendError(401)
             return;
-        } else if (!permissionHelperService.hasUserWithRole(result.user, result.institution, "INST_ADM")) {
+        } else if (!accessService.checkUserOrgRole(result.user, result.institution, "INST_ADM")) {
             flash.error = message(code: 'myinst.renewalUpload.error.noAdmin', default: 'Renewals Upload screen is not available to read-only users.')
             response.sendError(401)
             return;
@@ -2168,11 +2164,11 @@ AND EXISTS (
     def renewalsUpload() {
         def result = setResultGenerics()
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (! accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = message(code:'myinst.error.noMember', args:[result.institution.name]);
             response.sendError(401)
             return;
-        } else if (! permissionHelperService.hasUserWithRole(result.user, result.institution, "INST_ADM")) {
+        } else if (! accessService.checkUserOrgRole(result.user, result.institution, "INST_ADM")) {
             flash.error = message(code:'myinst.renewalUpload.error.noAdmin', default:'Renewals Upload screen is not available to read-only users.')
             response.sendError(401)
             return;
@@ -2334,7 +2330,7 @@ AND EXISTS (
         log.debug("-> renewalsUpload params: ${params}");
         def result = setResultGenerics()
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (! accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = message(code:'myinst.error.noMember', args:[result.institution.name]);
             response.sendError(401)
             // render(status: '401', text:"You do not have permission to access ${result.institution.name}. Please request access on the profile page");
@@ -2359,7 +2355,7 @@ AND EXISTS (
                 name: old_subname ?: "Unset: Generated by import",
                 startDate: sub_startDate,
                 endDate: sub_endDate,
-                instanceOf: old_subOID ?: null,
+                previousSubscription: old_subOID ?: null,
                 type: RefdataValue.findByValue('Subscription Taken'))
         log.debug("New Sub: ${new_subscription.startDate}  - ${new_subscription.endDate}")
         def packages_referenced = []
@@ -2523,14 +2519,14 @@ AND EXISTS (
     def dashboard() {
         def result = setResultGenerics()
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (! accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = "You do not have permission to access ${result.institution.name} pages. Please request access on the profile page";
             response.sendError(401)
             return;
         }
 
-        result.is_inst_admin = permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM')
-        result.editable = true // TODO: for tasks only
+        result.is_inst_admin = accessService.checkUserOrgRole(result.user, result.institution, 'INST_ADM')
+        result.editable = result.user?.hasAffiliation('INST_EDITOR')
 
         def pending_change_pending_status = RefdataCategory.lookupOrCreate("PendingChangeStatus", "Pending")
         getTodoForInst(result)
@@ -2589,7 +2585,7 @@ AND EXISTS (
     def changes() {
         def result = setResultGenerics()
 
-        if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+        if (! accessService.checkUserIsMember(result.user, result.institution)) {
             flash.error = "You do not have permission to view ${result.institution.name}. Please request access on the profile page";
             response.sendError(401)
             return;
@@ -2621,8 +2617,8 @@ AND EXISTS (
         result
     }
 
-    @DebugAnnotation(test = 'hasAffiliation("INST_USER")')
-    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_USER") })
+    @DebugAnnotation(test = 'hasAffiliation("INST_EDITOR")')
+    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_EDITOR") })
     def changeLog() {
         def result = setResultGenerics()
 
@@ -2692,12 +2688,12 @@ AND EXISTS (
         }
     }
 
-    @DebugAnnotation(test = 'hasAffiliation("INST_USER")')
-    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_USER") })
+    @DebugAnnotation(test = 'hasAffiliation("INST_EDITOR")')
+    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_EDITOR") })
     def financeImport() {
       def result = setResultGenerics()
 
-      if (! permissionHelperService.checkUserIsMember(result.user, result.institution)) {
+      if (! accessService.checkUserIsMember(result.user, result.institution)) {
           flash.error = "You do not have permission to view ${result.institution.name}. Please request access on the profile page";
           response.sendError(401)
           return;
@@ -2765,7 +2761,7 @@ AND EXISTS (
             }
         }
         result.visiblePersons = visiblePersons
-        result.editable = result.user.affiliations?.size() > 0
+        result.editable = accessService.checkMinUserOrgRole(result.user, contextService.getOrg(), 'INST_EDITOR') || SpringSecurityUtils.ifAnyGranted('ROLE_ADMIN')
 
         result
       }
@@ -2792,7 +2788,8 @@ AND EXISTS (
         def query = filterService.getTaskQuery(params, sdFormat)
         result.taskInstanceList   = taskService.getTasksByResponsibles(result.user, result.institution, query)
         result.myTaskInstanceList = taskService.getTasksByCreator(result.user, taskService.WITHOUT_TENANT_ONLY)
-        result.editable = true // TODO check roles !!!
+
+        result.editable = accessService.checkMinUserOrgRole(result.user, contextService.getOrg(), 'INST_EDITOR') || SpringSecurityUtils.ifAnyGranted('ROLE_ADMIN')
 
         log.debug(result.taskInstanceList)
         result
@@ -2801,8 +2798,8 @@ AND EXISTS (
     /**
      * Display and manage PrivateProperties for this institution
      */
-    @DebugAnnotation(test = 'hasAffiliation("INST_USER")')
-    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_USER") })
+    @DebugAnnotation(test = 'hasAffiliation("INST_EDITOR")')
+    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_EDITOR") })
     def managePrivateProperties() {
         def result = setResultGenerics()
 
@@ -2818,8 +2815,7 @@ AND EXISTS (
         result
     }
 
-    @DebugAnnotation(test = 'hasAffiliation("INST_USER")')
-    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_USER") })
+    @Secured(['ROLE_USER'])
     def switchContext() {
         def user = User.get(springSecurityService.principal.id)
         def org  = genericOIDService.resolveOID(params.oid)
@@ -2918,7 +2914,7 @@ AND EXISTS (
         def result = setResultGenerics()
         result.orgType = result.institution.orgType
 
-        result.editable = permissionHelperService.hasUserWithRole(result.user, result.institution, 'INST_ADM')
+        result.editable = accessService.checkUserOrgRole(result.user, result.institution, 'INST_ADM')
         if (result.editable) {
 
             if(result.orgType?.value == 'Consortium') {
@@ -2929,5 +2925,33 @@ AND EXISTS (
             result
         }
         render (template: "../templates/filter/orgFilterTable", model: [orgList: result.cons_members, tmplShowCheckbox: true])
+    }
+
+    @DebugAnnotation(test='hasAffiliation("INST_ADM")')
+    @Secured(closure = { ctx.springSecurityService.getCurrentUser()?.hasAffiliation("INST_ADM") })
+    def actionAffiliationRequestOrg() {
+        log.debug("actionMembershipRequestOrg");
+        def req = UserOrg.get(params.req);
+        def user = User.get(springSecurityService.principal.id)
+        def currentOrg = contextService.getOrg()
+        if ( req != null && req.org == currentOrg) {
+            switch(params.act) {
+                case 'approve':
+                    req.status = UserOrg.STATUS_APPROVED
+                    break;
+                case 'deny':
+                    req.status = UserOrg.STATUS_REJECTED
+                    break;
+                default:
+                    log.error("FLASH UNKNOWN CODE");
+                    break;
+            }
+            req.dateActioned = System.currentTimeMillis();
+            req.save(flush:true);
+        }
+        else {
+            log.error("FLASH");
+        }
+        redirect(action: "index")
     }
 }